feat: farmer detail page, fix null specializations
