+++ apps/client-fnp/.env.production
@@ -1,5 +1,5 @@
 NEXT_PUBLIC_APP_URL=https://farmnport.com
 NEXT_PUBLIC_BASE_URL=https://farmnport.com
-NEXT_PUBLIC_GTM_ID=GTM-KBM2DJ86
+NEXT_PUBLIC_GTM_ID=GTM-TS7XJ4J
 NEXT_TELEMETRY_DEBUG=0
 NEXTAUTH_URL=https://farmnport.com
